Исправляет конструктор класса и геттер шаблона точки маршрута
