--- src/create-element.js
@@ -1,5 +1,0 @@
-export const createElement = (template) => {
-  const newElement = document.createElement(`template`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
